+++ src/cucu/browser/core.py
@@ -14,64 +14,64 @@ class Browser:
         pass
 
     def open(self, *args, **kwargs):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def navigate(self, url):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_next_tab(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_previous_tab(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def back(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def refresh(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def get_current_url(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def title(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def execute(self, javascript, *args, **kwargs):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def click(self, element):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_default_frame(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_frame(self, frame):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def screenshot(self, filepath):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def close_window(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def get_tab_info(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def get_all_tabs_info(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_nth_tab(self, tab_index):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def switch_to_tab_that_matches_regex(self, text):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def get_session_id(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     def quit(self):
-        raise RuntimeError("implement me")
+        raise AssertionError("implement me")
 
     # built in methods to be used by all browser implementations
     def wait_for_page_to_load(self):