+++ src/cucu/steps/dropdown_steps.py
@@ -142,7 +142,7 @@ def click_dropdown(ctx, dropdown):
     ctx.check_browser_initialized()
 
     if base_steps.is_disabled(dropdown):
-        raise RuntimeError("unable to click the button, as it is disabled")
+        raise AssertionError("unable to click the button, as it is disabled")
 
     logger.debug("clicking dropdown")
     try:
@@ -188,10 +188,10 @@ def find_n_select_dropdown_option(ctx, dropdown, option, index=0):
 
     if dropdown_element is None:
         prefix = "" if index == 0 else f"{humanize.ordinal(index + 1)} "
-        raise RuntimeError(f"unable to find the {prefix}dropdown {dropdown}")
+        raise AssertionError(f"unable to find the {prefix}dropdown {dropdown}")
 
     if base_steps.is_disabled(dropdown_element):
-        raise RuntimeError(
+        raise AssertionError(
             "unable to select from the dropdown, as it is disabled"
         )
 
@@ -207,7 +207,7 @@ def find_n_select_dropdown_option(ctx, dropdown, option, index=0):
         option_element = find_dropdown_option(ctx, option)
 
         if option_element is None:
-            raise RuntimeError(
+            raise AssertionError(
                 f'unable to find option "{option}" in dropdown "{dropdown}" for selection'
             )
 
@@ -232,10 +232,10 @@ def find_n_select_dynamic_dropdown_option(ctx, dropdown, option, index=0):
 
     if dropdown_element is None:
         prefix = "" if index == 0 else f"{humanize.ordinal(index + 1)} "
-        raise RuntimeError(f"unable to find the {prefix}dropdown {dropdown}")
+        raise AssertionError(f"unable to find the {prefix}dropdown {dropdown}")
 
     if base_steps.is_disabled(dropdown_element):
-        raise RuntimeError(
+        raise AssertionError(
             "unable to select from the dropdown, as it is disabled"
         )
 
@@ -273,7 +273,7 @@ def find_n_select_dynamic_dropdown_option(ctx, dropdown, option, index=0):
                 option_element = None
 
     if option_element is None:
-        raise RuntimeError(
+        raise AssertionError(
             f'unable to find option "{option}" in dropdown "{dropdown}" for dynamic selection'
         )
 
@@ -297,7 +297,7 @@ def assert_dropdown_option_selected(
 
     dropdown_element = find_dropdown(ctx, dropdown, index)
     if dropdown_element is None:
-        raise RuntimeError(f'unable to find dropdown "{dropdown}"')
+        raise AssertionError(f'unable to find dropdown "{dropdown}"')
 
     selected_option = None
     if dropdown_element.tag_name == "select":
@@ -305,7 +305,7 @@ def assert_dropdown_option_selected(
         selected_option = select_element.first_selected_option
 
         if selected_option is None:
-            raise RuntimeError(
+            raise AssertionError(
                 f"unable to find selected option in dropdown {dropdown}"
             )
 
@@ -316,10 +316,10 @@ def assert_dropdown_option_selected(
         #      and easier to find for automation by their name
         if is_selected:
             if selected_name.find(option) == -1:
-                raise RuntimeError(f"{option} is not selected")
+                raise AssertionError(f"{option} is not selected")
         else:
             if selected_name.find(option) != -1:
-                raise RuntimeError(f"{option} is selected")
+                raise AssertionError(f"{option} is selected")
 
     else:
         if dropdown_element.get_attribute("aria-expanded") != "true":
@@ -329,16 +329,16 @@ def assert_dropdown_option_selected(
         selected_option = find_dropdown_option(ctx, option)
 
         if selected_option is None:
-            raise RuntimeError(
+            raise AssertionError(
                 f'unable to find option "{option}" in dropdown "{dropdown}" for assertion'
             )
 
         if is_selected:
             if selected_option.get_attribute("aria-selected") != "true":
-                raise RuntimeError(f"{option} is not selected")
+                raise AssertionError(f"{option} is not selected")
         else:
             if selected_option.get_attribute("aria-selected") == "true":
-                raise RuntimeError(f"{option} is selected")
+                raise AssertionError(f"{option} is selected")
 
         # close the dropdown
         click_dropdown(ctx, dropdown_element)