+++ tests/test_browser.py
@@ -14,7 +14,7 @@ def test_any_subclass_of_browser_gets_appropriate_error_from_unimplemented_class
         if method_name in ignore_methods:
             continue
 
-        with pytest.raises(RuntimeError, match="implement me"):
+        with pytest.raises(AssertionError, match="implement me"):
             method = browser.__getattribute__(method_name)
             print(method.__code__.co_argcount)
             args = [